ANDROID: usb: host: xhci-mem: Cleanup pending secondary event ring events
As part of xHCI bus suspend, the XHCI is halted.  However, if there are
pending events in the secondary event ring, it is observed that the xHCI
controller stops responding to further commands upon host or device
initiated bus resume.  Iterate through all pending events and update the
dequeue pointer to the beginning of the event ring.

Bug: 264371591
CRs-Fixed: 3651753
Change-Id: Ida0299d9d37e804658f0ca627215255b81d9f99b
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>