ANDROID: usb: host: xhci: Fixup XHCI changes to allow XHCI sideband as module
Currently, the XHCI sideband driver is compiled as part of the GKI owned
XHCI HCD object.  However, this is not possible, since the GKI image is
built on configurations defined in gki_defconfig.  Modify the Kconfig and
Makefile for this driver to allow for compilation as a vendor module.

Bug: 264371591
Bug: 319144906
CRs-Fixed: 3651753
Change-Id: I4184010d44d6cc638c403cc6bf36fd5635bc840e
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>
(cherry picked from commit 7f2e17fbe48d5c766f58ee870f1d7962ffc27594)